Day 16 Part 1
It was taking too long to visit all the options even with the trimming function
After unleashing Part 2, the path is not really needed, so I could get the graph
and simplify it by just using the valves with flow_rate>0 and getting the distances
from valve to valve

The trimming function is still useful. It turns a 5s solution into a ms one
visit and open are only needed for the tests

Orienting the solution to be fully testable has been a mistake here.
The sample is one of the several paths that get the optimal route
I was not able to infer a strategy from the sample. DFS used to check them all